fix: move logic from plugin to skybox controller
