+++ Explorer/Assets/DCL/StylizedSkybox/Scripts/Plugin/StylizedSkyboxPlugin.cs
@@ -50,47 +50,13 @@ namespace DCL.StylizedSkybox.Scripts.Plugin
             skyboxController = Object.Instantiate((await assetsProvisioner.ProvideMainAssetAsync(settingsAsset.StylizedSkyboxPrefab, ct: ct)).Value.GetComponent<SkyboxController>());
             AnimationClip skyboxAnimation = (await assetsProvisioner.ProvideMainAssetAsync(settingsAsset.SkyboxAnimationCycle, ct: ct)).Value;
 
-            skyboxController.Initialize(settingsAsset.SkyboxMaterial, directionalLight, skyboxAnimation, featureFlagsCache);
-
-            settingsAsset.NormalizedTime = skyboxController!.DynamicTimeNormalized;
-            settingsAsset.NormalizedTimeChanged += OnNormalizedTimeChanged;
-            settingsAsset.UseDynamicTime = skyboxController.UseDynamicTime;
-            settingsAsset.UseDynamicTimeChanged += OnUseDynamicTimeChanged;
-
-            skyboxController.OnSkyboxUpdated += OnSkyboxUpdated;
+            skyboxController.Initialize(settingsAsset.SkyboxMaterial, directionalLight, skyboxAnimation, featureFlagsCache, settingsAsset);
 
             debugContainerBuilder.TryAddWidget("Skybox")
                                 ?.AddSingleButton("Play", () => skyboxController.UseDynamicTime = true)
                                  .AddSingleButton("Pause", () => skyboxController.UseDynamicTime = false)
                                  .AddFloatSliderField("Time", timeOfDay, 0, 1)
                                  .AddSingleButton("SetTime", () => skyboxController.SetTimeOverride(timeOfDay.Value)); //TODO: replace this by a system to update the value
-        }
-
-        private void OnSkyboxUpdated()
-        {
-            // When skybox gets dynamically updated we refresh the
-            // settings value so it reflects the current state
-
-            if (skyboxController!.UseDynamicTime)
-            {
-                settingsAsset!.NormalizedTime = skyboxController.DynamicTimeNormalized;
-            }
-        }
-
-        private void OnUseDynamicTimeChanged(bool dynamic)
-        {
-            skyboxController!.UseDynamicTime = dynamic;
-
-            if (dynamic)
-                settingsAsset!.NormalizedTime = skyboxController!.DynamicTimeNormalized;
-        }
-
-        private void OnNormalizedTimeChanged(float tod)
-        {
-            if (!skyboxController!.UseDynamicTime) // Ignore updates to the value when they come from the skybox
-            {
-                skyboxController!.SetTimeOverride(tod);
-            }
         }
 
         [Serializable]
